fix up pricing formatting in listings csv and implement op codes for mint splits

+++ operator_cli/commands/storefront/types.go
@@ -1,11 +1,12 @@
 package storefront
 
 type Catalog struct {
-	BatchAccount   string `json:"batchAccount"`
-	ConfigIndex    int    `json:"configIndex"`
-	IsListed       bool   `json:"isListed"`
-	Price          int    `json:"price"`
-	LifecycleStart int    `json:"lifecycleStart"`
-	Mints          int    `json:"mints"`
-	Resync         bool   `json:"resync"`
+	BatchAccount   string  `json:"batchAccount"`
+	ConfigIndex    int     `json:"configIndex"`
+	IsListed       bool    `json:"isListed"`
+	Price          float64 `json:"price"`
+	LifecycleStart int     `json:"lifecycleStart"`
+	Mints          int     `json:"mints"`
+	Resync         bool    `json:"resync"`
 }
+
